support version overwrite

+++ src/commands/publish.ts
@@ -5,6 +5,7 @@ import {zip} from 'zip-a-folder'
 
 import {loadConfig, parseConfigOverwrites} from '../utils'
 import {BackendFactory} from '../backends/factory'
+import {ModuleAlreadyExistsError} from '../errors'
 
 export default class Publish extends Command {
   static description = 'Publish a terraform module'
@@ -14,6 +15,10 @@ export default class Publish extends Command {
   ]
 
   static flags = {
+    overwrite: Flags.boolean({
+      summary: 'Overwrite a published version with new artifact',
+      default: false,
+    }),
     'overwrite-config': Flags.string({
       summary: 'Overwrite terrac configuration',
       multiple: true,
@@ -28,14 +33,6 @@ export default class Publish extends Command {
     const {flags} = await this.parse(Publish)
 
     const workDir = flags['work-directory']
-
-    // const git = simpleGit(workDir)
-    // const gitStatus = await git.status()
-
-    // if (gitStatus.not_added.length > 0 || gitStatus.staged.length > 0) {
-    //   throw new Error('workspace not clean')
-    // }
-
     const config = await loadConfig(workDir, parseConfigOverwrites(flags['overwrite-config']))
 
     const zipPath = `${tmpdir()}/${config.module.name}-${config.module.version}.zip`
@@ -44,6 +41,10 @@ export default class Publish extends Command {
     const backend = BackendFactory.create(config.backend)
     const name = config.module.name
     const version = config.module.version
+
+    if ((await backend.exists(name, version)) && !flags.overwrite) {
+      throw new ModuleAlreadyExistsError(name, version)
+    }
 
     await backend.publish(name, version, zipPath)
     await unlink(zipPath)
